Utils: Add @gui_require decorator
