+++ orangecontrib/text/widgets/utils/__init__.py
@@ -1,2 +1,3 @@
+from .decorators import *
 from .widgets import *
 from .concurrent import OWConcurrentWidget, asynchronous
